update share menu
