+++ apps/dotcom/client/src/tla/components/TlaFileShareMenu/file-share-menu.module.css
@@ -1,0 +1,114 @@
+/* ------------------- Share menu ------------------- */
+
+.shareMenu {
+	width: 240px;
+	--share-menu-padding: 12px;
+}
+
+.content {
+	position: relative;
+}
+
+.section {
+	display: flex;
+	padding: var(--share-menu-padding);
+	flex-direction: column;
+	gap: var(--share-menu-padding);
+}
+
+.section:nth-of-type(n + 2) {
+	border-top: 1px solid var(--tla-color-border);
+}
+
+.controlGroup {
+	display: flex;
+	flex-direction: column;
+	gap: 4px;
+}
+
+.control {
+	display: flex;
+	align-items: center;
+	justify-content: space-between;
+	height: 32px;
+	gap: 8px;
+}
+
+.controlButton {
+	all: unset;
+	width: 100%;
+	height: 100%;
+	display: flex;
+	align-items: center;
+	justify-content: space-between;
+	cursor: pointer;
+}
+
+/* Copy button */
+
+.copyButtonIcon {
+	position: absolute;
+	right: 10px;
+}
+
+/* QR Code */
+
+.qrCode {
+	position: relative;
+	width: 100%;
+	height: 216px;
+	padding: 0px var(--share-menu-padding);
+	margin: 0px 0px var(--share-menu-padding) 0px;
+}
+
+.qrCodeInner {
+	width: 100%;
+	height: 100%;
+	background-color: var(--tla-color-low);
+	background-size: 100% 100%;
+	background-position: center center;
+	background-color: var(--tla-color-low);
+}
+
+/* Share menu description */
+
+.description {
+	color: var(--tla-color-text-1);
+	font-weight: 400;
+	background-color: var(--tla-color-low);
+	border-radius: 6px;
+	font-size: 12px;
+	font-weight: 400;
+	line-height: 16px;
+	display: flex;
+	flex-direction: column;
+}
+
+.description p {
+	margin: 0px;
+	padding: 12px 12px 0px 12px;
+}
+
+.descriptionDismiss {
+	all: unset;
+	width: fit-content;
+	align-self: flex-end;
+	background: none;
+	border: none;
+	padding: 12px;
+	color: var(--tla-color-primary);
+	font-weight: 500;
+	text-align: right;
+	cursor: pointer;
+}
+
+.descriptionCollapsed {
+	all: unset;
+	display: flex;
+	align-items: center;
+	justify-content: center;
+	gap: 4px;
+	font-size: 11px;
+	text-align: center;
+	color: var(--tla-color-primary);
+}
