+++ apps/dotcom/client/src/tla/components/TlaSelect/select.module.css
@@ -1,0 +1,20 @@
+.container {
+	position: relative;
+}
+
+.select {
+	position: absolute;
+	top: 0px;
+	left: 0px;
+	width: 100%;
+	height: 100%;
+	opacity: 0;
+	cursor: pointer;
+	pointer-events: all;
+}
+
+.label {
+	display: flex;
+	flex-direction: row;
+	gap: 4px;
+}
