+++ apps/dotcom/client/src/tla/pages/file.tsx
@@ -3,7 +3,7 @@ import { useParams } from 'react-router-dom'
 import { useValue } from 'tldraw'
 import { TlaEditor } from '../components/TlaEditor'
 import { TlaErrorPage } from '../components/TlaErrorPage'
-import { TlaFileShareMenu } from '../components/TlaFileShareMenu'
+import { TlaFileShareMenu } from '../components/TlaFileShareMenu/TlaFileShareMenu'
 import { TlaSidebarToggle } from '../components/TlaSidebarToggle'
 import { TlaWrapperWithSidebar } from '../components/TlaWrapperWithSidebar'
 import { useApp } from '../hooks/useAppState'
