+++ apps/dotcom/client/src/tla/styles/tla.css
@@ -336,42 +336,6 @@
 	}
 }
 
-/* ------------------ Toggle switch ----------------- */
-
-.tla-switch {
-	--switch-h: 24px;
-	--switch-w: calc(var(--switch-h) * 1.618);
-	height: var(--switch-h);
-	width: var(--switch-w);
-	position: relative;
-	border-radius: 99px;
-	background-color: var(--tla-color-text-3);
-	transition: all 0.12s;
-}
-
-.tla-switch:after {
-	display: block;
-	content: '';
-	position: absolute;
-	top: 2px;
-	left: 2px;
-	right: initial;
-	border-radius: 100%;
-	height: calc(var(--switch-h) - 4px);
-	width: calc(var(--switch-h) - 4px);
-	background-color: var(--tla-color-panel);
-	pointer-events: none;
-	transition: all 0.12s;
-}
-
-.tla-switch[data-checked='true'] {
-	background-color: var(--tla-color-primary);
-}
-
-.tla-switch[data-checked='true']:after {
-	left: calc(var(--switch-w) - (var(--switch-h) - 4px) - 2px);
-}
-
 /* --------------------- Inputs --------------------- */
 
 .tla-input {
@@ -1128,128 +1092,4 @@ want centered content that docks to the top when the window is short
 
 .tla-logged-out__footer a {
 	color: var(--tla-color-primary);
-}
-
-/* ------------------- Share menu ------------------- */
-
-.tla-share-menu {
-	width: 240px;
-	--share-menu-padding: 12px;
-}
-
-.tla-share-menu__content {
-	padding: 10px 0px var(--share-menu-padding) 0px;
-}
-
-/* Share toggle */
-
-.tla-share-menu__control {
-	display: flex;
-	align-items: center;
-	justify-content: space-between;
-	padding-left: var(--share-menu-padding);
-	height: 40px;
-}
-
-.tla-share-menu__control-button {
-	all: unset;
-	width: 100%;
-	height: 100%;
-	display: flex;
-	align-items: center;
-	justify-content: space-between;
-	cursor: pointer;
-}
-
-.tla-share-menu__control-container {
-	position: relative;
-	height: 100%;
-	display: flex;
-	align-items: center;
-	justify-content: center;
-	padding: 0px var(--share-menu-padding);
-	cursor: pointer;
-	pointer-events: all;
-}
-
-.tla-share-menu__control-container select,
-.tla-share-menu__control-container input[type='checkbox'] {
-	position: absolute;
-	top: 0px;
-	left: 0px;
-	width: 100%;
-	height: 100%;
-	opacity: 0;
-	cursor: pointer;
-	pointer-events: all;
-}
-
-/* Select */
-
-.tla-share-menu__select__label {
-	display: flex;
-	flex-direction: row;
-	gap: 4px;
-}
-
-/* Copy button */
-
-.tla-share-menu__copy-link {
-	padding: 0px var(--share-menu-padding);
-	width: 100%;
-}
-
-.tla-share-menu__copy-button {
-	width: 100%;
-	margin: calc(var(--share-menu-padding) / 2) 0px;
-}
-
-.tla-share-menu__copy-button > .tla-icon {
-	position: absolute;
-	right: 10px;
-}
-
-.tla-share-menu .tla-divider {
-	margin: var(--share-menu-padding) 0px;
-}
-
-/* QR Code */
-
-.tla-share-menu__qr-code {
-	position: relative;
-	width: 100%;
-	height: 216px;
-	padding: 0px var(--share-menu-padding);
-	margin: 0px 0px var(--share-menu-padding) 0px;
-}
-
-.tla-share-menu__qr-code-inner {
-	width: 100%;
-	height: 100%;
-	background-color: var(--tla-color-low);
-	background-size: 100% 100%;
-	background-position: center center;
-	background-color: var(--tla-color-low);
-}
-
-/* Share menu description */
-
-.tla-share-menu__description {
-	padding: 12px 0px;
-	color: var(--tla-color-text-1);
-	font-weight: 400;
-	background-color: var(--tla-color-low);
-	padding: 12px 12px 0px 12px;
-	border-radius: 6px;
-}
-
-.tla-share-menu__description-dismiss {
-	background: none;
-	border: none;
-	padding: 12px 0px;
-	color: var(--tla-color-primary);
-	font-weight: 500;
-	width: 100%;
-	text-align: right;
-	cursor: pointer;
 }
